Move logs to their own directory /var/log/webdavd
This is more in line with current practice for services that create
their own logfiles.

Add a line to the systemd unit that will create the log directory before
starting.  systemd will also handle permissions on the directory and can
do various things with namespaces to prevent access to other files in
/var or elsewhere.

Update logrotate to match.

Remove owner/permission setting for logrotate.  It will copy the
existing logfiles' values, which allows it to work properly if webdavd
runs as another user or with a different umask.