Remove lock attempts on GET and PROPFIND
Neither of these requests should block if the file has a write lock.

Clients will not provide a lock id in the requests header for these
requests, even if they have lack.  It appears to be standard practice
for clients writing to files to lock them and then make GET and/on
PROPFIND calls while they are locked.

The result is webdavd hangs with most clients that support writing.

Removing the lock attempts for GET and PROPFIND makes them work.

It's possible that *read* locks should interact with GET/PROPFIND.  The
RFC says almost nothing about how read locks should work and appears to
give the implementation free reign to define their semantics.

In practice, nobody uses read locks, so it doesn't matter.