Add support for socket based activation
This allows using with systemd socket units.  It works like the old
inetd style activation.

The socket(s) to open are described in systemd socket units, and when a
connection is made, systemd starts webdavd and passes the sockets to it
on some file descriptors.

This way webdavd doesn't need to be running until it's used.  Systemd
also supports a lot more options for how it can listen on a socket.
E.g., sockets can be bound to a specific interface and UNIX domain
sockets can be used.

Multiple sockets can be used, e.g. http and https.  Systemd will pass
all of them to webdavd.

libminihttpd supports this already and can take an existing fd as the
socket.

I use a systemd function that does some handy stuff, so this requires
systemd's libraries.  So that this isn't a hard dependency, it's
protected in the build system by requiring HAVE_SYSTEMD to be set at
build time, e.g. make HAVE_SYSTEMD=1.

One could add support for the old inetd server, but I didn't since it's
not been included with with Fedora for years.