Use MHD_OPTION_ARRAY to pass extra options to MHD_start_daemon
The call to MHD_start_daemon() was duplicated twice, with slightly
different options, for ssl vs non-ssl.

Use MHD_OPTION_ARRAY so that the same call can be used for both with the
ssl option passed via the array.

This removes some duplicated code.  Right now it is not much, because
ssl is the only option.  It will make a much difference when there is
another option.  The current method will require a different call for
every possible combination of options.