Add socket units and package with RPM
This enables the systemd socket activation support when building the
RPM.

It includes two sample socket unit files.  They will start webdavd if
there is a connection on either the http or https port.

They can be used by starting/enabling the socket units instead of the
service unit.  It's possible to enable either or both of the socket
units.

and also