Continue when socket can't be used
This doesn't really do much of anything.  Mostly it changes the flow so
a socket bind address failure continues instead having the rest of
the loop inside an if statement.

So it makes a bunch of code indented one level fewer.

The real reason is so the next commit, when the socket address might not
be used, doesn't need to do this.  That way it doesn't look like a lot
of code is being re-written when it's not.