Fix hangs when reading bad xml
The xml library functions return -1 on error and 0 on EOF.  A number of
loops of the form "while (result)" would loop forever on an error.

Change this to "while (result > 0)", so that they exit on an error.

The return value of the functions in xml.c that did this will be 1 on
success and 0 on EOF *or error*.  So the code using the xml.c functions
will be ok as is.

Example that hangs:
<listen> <port> 123 <port> </listen>

The 2nd <port> is missing the slash and doesn't terminate the element,
so </listen> is unmatched and it will hang on that error.